msm: camera: Fix camera cannot open
The driver structure is reset just right after probe successful.
Change the logical to reset s_ctrl only when probe failed.

Test: Build pass, camera is working
Bug: 115421538
Change-Id: I596f4cbc1c39a1e49ddfbee1db1610fd73a6da00
Signed-off-by: Speth Chang <spethchang@google.com>

[Ayrton]: After checking the new logs generated by the commit 040be97
I found out that camera was closing the lenses by it's own causing
camera app to crash, so according to the info, this commit should fix
the camera issues with kernel on custom roms.

Signed-off-by: Official-Ayrton990 <15030201@itcelaya.edu.mx>